trading-bot/requirements.txt
python-telegram-bot==13.5
schedule==1.1.0
sqlite3
requests==2.25.1
web3==5.17.0
solana==0.11.3
